+++ analysis_options.yaml
@@ -107,7 +107,7 @@ linter:
 #    - unnecessary_parenthesis # under review (see #1068)
 #    - unnecessary_statements # under review (see #1068)
     - unnecessary_this
-    - unsafe_script_src
+    - unsafe_html
     - use_setters_to_change_properties
     - use_string_buffers
     - use_to_and_as_if_applicable
